login/logout and user restrictions done

+++ project/base/views.py
@@ -1,7 +1,13 @@
 from .forms import *
 from django.shortcuts import render, redirect
+from django.contrib import messages
 from .models import *
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+
 # Create your views here.
 
 def home(request):
@@ -44,6 +50,7 @@ def room(request,pk):
 
     return render(request,"room.html", context)
 
+@login_required(login_url='login')
 def createRoom(request):
 
     form = RoomForm()
@@ -60,11 +67,14 @@ def createRoom(request):
 
     return render(request, "room_form.html", context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     
     room = Room.objects.get(id=pk)
-
     form = RoomForm(instance=room) #instance koymamızın nedeni id=pk olanı çevirmek ve onu editlemek
+    
+    if request.user != room.host: # bu olay çok önemli
+            return HttpResponse('Your are not allowed here!!')
 
     if request.method == "POST":
         form = RoomForm(request.POST,instance=room)
@@ -78,14 +88,48 @@ def updateRoom(request, pk):
 
     return render(request, "room_form.html", context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
 
     # if request.user != room.host:
     #     return HttpResponse('Your are not allowed here!!')
 
+    if request.user != room.host: # bu olay çok önemli
+            return HttpResponse('Your are not allowed here!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
 
     return render(request, 'delete.html', {'obj': room})
+
+def loginPage(request):
+    # page = 'login'
+    # if request.user.is_authenticated:
+    #      return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+        #here we are checking if user exist. If yes, username i al, if not, error mesaj dön
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username OR password does not exist')
+
+    context = {}
+    return render(request, 'login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
